feat: watched-page에서 삭제, view, like 기능
