+++ src/js/controller/watched-page.ts
@@ -1,5 +1,6 @@
 import { $ } from '../@shared/utils/utils.js';
 import { renderWatchedPage } from '../view/watched-page.js';
+import { clickCheckButtons } from './index.js';
 
 const clickWatchedPageButton = () => {
   renderWatchedPage();
@@ -7,6 +8,7 @@ const clickWatchedPageButton = () => {
 
 const watchedPageController = () => {
   $('#watched-page-button')?.addEventListener('click', clickWatchedPageButton);
+  $('#main-videos')?.addEventListener('click', clickCheckButtons);
 };
 
 export { watchedPageController };
